requirements.txt
rembg
pillow
pyheif_pillow_opener
io